load-generator-instabook/requirements.txt
requests==2.26.0
prometheus-client==0.11.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-exporter-otlp==1.20.0
